requirements.txt
Flask
pytesseract
pdf2image
camelot-py[cv]
python-docx
PyPDF2
pandas
gunicorn
